Cast (size_t) to (uint32_t) when printing.
git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@285 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/bgp/notification_packet.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/bgp/notification_packet.cc,v 1.10 2003/01/26 04:06:17 pavlin Exp $"
+#ident "$XORP: xorp/bgp/notification_packet.cc,v 1.11 2003/01/29 20:32:31 rizzo Exp $"
 
 #include "bgp_module.h"
 #include "config.h"
@@ -77,7 +77,7 @@ NotificationPacket::NotificationPacket(const uint8_t *d, uint16_t l)
 const uint8_t *
 NotificationPacket::encode(size_t& len, uint8_t *buf) const
 {
-    debug_msg("Encode in NotificationPacket called (%d)\n", _Length);
+    debug_msg("Encode in NotificationPacket called (%u)\n", (uint32_t)_Length);
     if (buf != 0)		// have a buffer, check length
 	assert(len >= _Length);	// XXX should become an exception
